requirements.txt
beautifulsoup4==4.10.0
demjson==1.6
lxml==4.7.1
pycryptodome==3.12.0
PyMySQL==1.0.2
PyQt5==5.15.6
PySide2==5.15.2
python_bcrypt==0.3.2
scrypt==0.8.18
